a68: modules exports
This commit adds the code that handles the exports information for the
module definitions in prelude packets.  The exports info is generated
in a section in the output object file.

A precise description of the binary format in which the exports are
encoded is expressed in an included GNU poke pickle ga68-exports.pk.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

gcc/ChangeLog

	* algol68/a68-exports.cc: New file.
	* algol68/ga68-exports.pk: Likewise.