c++: Fix up [[maybe_unused]] handling on expansion stmts [PR122788]
This PR complains that [[maybe_unused]] attribute is ignored on
the range-for-declaration of expansion-statement.

We copy DECL_ATTRIBUTES and apply late attributes, but early attributes
don't have their handlers called again, so some extra flags need to be
copied as well.
This copies TREE_USED and DECL_READ_P flags.

2025-11-22  Jakub Jelinek  <jakub@redhat.com>

	PR c++/122788
	* pt.cc (finish_expansion_stmt): Or in TREE_USED and DECL_READ_P
	flags from range_decl to decl or from corresponding structured binding
	to this_decl.

	* g++.dg/cpp26/expansion-stmt27.C: New test.