OpenMP: Fix "begin declare variant" test failure with -m32
As reported by Haochen Jiang, this recently-added test case was
failing on x86_64 with -m32; the target hook for matching the "arch"
selector won't match "x86_64" in that case, even if gcc was configured
for that target.  It does match plain "x86" for both 64 and 32 bit targets,
so I've switched the testcase to use that instead.

Committed as obvious (at least in retrospect).

gcc/testsuite/ChangeLog
	* c-c++-common/gomp/delim-declare-variant-6.c (f3): Use "x86"
	instead of "x86_64" in the arch selector, to match both 64- and
	32-bit targets.