Adjust gcc.dg/vect/bb-slp-41.c
We now perform SLP vectorization for the scalar epilog of a vector
loop where we now unroll said scalar epilog.  The following adjusts
the testcase to look for SLP trying to vectorize a CTOR instead,
which I guessed from the topic of r10-4336-g818b3293f4545d which
added this testcase.

	* gcc.dg/vect/bb-slp-41.c: Adjust dump scan.