phiprop: Small compile time improvement for phiprop
Now that post dom information is only needed when the new store
can trap (since r16-5555-g952e145796d), only calculate it when
that is the case. It was calculated on demand by
r14-2051-g3124bfb14c0bdc. This just changes when we need to
calculate it.

Pushed as obvious.

gcc/ChangeLog:

	* tree-ssa-phiprop.cc (propagate_with_phi): Only
	calculate on demand post dom info when the new store
	might trap.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>