phiprop: Make sure types of the load match the inserted phi [PR122847]
This was introduced with r16-5556-ge94e91d6f3775, but the type
check for the delay was not happen because the type at the point
of delaying was set to NULL. It is only until a non-delayed load
when the phi is created, the type is set.

This adds the type check to the replacement for the delayed statements.

Pushed as obvious.

	PR tree-optimization/122847

gcc/ChangeLog:

	* tree-ssa-phiprop.cc (propagate_with_phi): Add type
	check for reuse of the phi for the delayed statements.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr122847-1.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>